app/requirements.txt
-i https://pypi.org/simple
anyio==4.3.0; python_version >= '3.8'
ariadne==0.23.0
asn1crypto==1.5.1
boto3==1.34.105; python_version >= '3.8'
botocore==1.34.105; python_version >= '3.8'
fastapi==0.99.0; python_version >= '3.7'
graphql-core==3.2.3; python_version >= '3.6' and python_version < '4'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.7; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
mangum==0.17.0; python_version >= '3.7'
pg8000==1.31.2; python_version >= '3.8'
pydantic==1.10.15; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.8
s3transfer==0.10.1; python_version >= '3.8'
scramp==1.4.5; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.30; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
typing-extensions==4.11.0; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.10'
